pyproject.toml
[project]
name = "telegram2foam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
]
